output: Extract file review footer hints
File-review footer rendering built command hints and printed the footer in
one module. Hint construction owns review-state shims, line action
commands, navigation commands, and selection formatting.

Keeping that logic beside terminal styling made the footer renderer depend
on command policy and selected-change record details.

This commit moves hint construction into output.file_review_footer_hints
and leaves output.file_review_footer to style and print the returned hints.

Subsequent commits will add architecture coverage for footer hint
ownership.
